+++ example/schema-relay-paginate-2/QueryRoot.js
@@ -1,0 +1,80 @@
+import {
+  GraphQLObjectType,
+  GraphQLList,
+  GraphQLString,
+  GraphQLInt
+} from 'graphql'
+
+import { connectionArgs } from 'graphql-relay'
+
+import knex from './database'
+import { User, UserConnection } from './User'
+import Sponsor from './Sponsor'
+import { nodeField } from './Node'
+
+import joinMonster from '../../src/index'
+const options = {
+  minify: process.env.MINIFY == 1,
+  dialect: process.env.PG_URL ? 'pg' : 'standard'
+}
+
+export default new GraphQLObjectType({
+  description: 'global query object',
+  name: 'Query',
+  fields: () => ({
+    version: {
+      type: GraphQLString,
+      resolve: () => joinMonster.version
+    },
+    node: nodeField,
+    users: {
+      type: UserConnection,
+      args: connectionArgs,
+      sqlPaginate: true,
+      sortKey: {
+        order: 'asc',
+        key: 'id'
+      },
+      resolve: (parent, args, context, ast) => {
+        return joinMonster(ast, context, sql => {
+          // place the SQL query in the response headers. ONLY for debugging. Don't do this in production
+          if (context) {
+            context.set('X-SQL-Preview', sql.replace(/\n/g, '%0A'))
+          }
+          return knex.raw(sql)
+        }, options)
+      }
+    },
+    user: {
+      type: User,
+      args: {
+        id: {
+          description: 'The users ID number',
+          type: GraphQLInt
+        }
+      },
+      where: (usersTable, args, context) => { // eslint-disable-line no-unused-vars
+        if (args.id) return `${usersTable}.id = ${args.id}`
+      },
+      resolve: (parent, args, context, ast) => {
+        return joinMonster(ast, context, sql => {
+          if (context) {
+            context.set('X-SQL-Preview', sql.replace(/\n/g, '%0A'))
+          }
+          return knex.raw(sql)
+        }, options)
+      }
+    },
+    sponsors: {
+      type: new GraphQLList(Sponsor),
+      resolve: (parent, args, context, ast) => {
+        // use the callback version this time
+        return joinMonster(ast, context, (sql, done) => {
+          knex.raw(sql)
+          .then(data => done(null, data))
+          .catch(done)
+        }, options)
+      }
+    }
+  })
+})