+++ src/util.js
@@ -9,20 +9,31 @@ export function inspect(obj, options = {}) {
   return util.inspect(obj, { depth: 12, ...options })
 }
 
+// really? yes, really
+export function last(arr) {
+  return arr[arr.length - 1]
+}
+
+export function wrap(maybeArr) {
+  if (maybeArr.constructor === Array) {
+    return maybeArr
+  }
+  return [ maybeArr ]
+}
+
 export function validateSqlAST(topNode) {
+  // TODO: this could be a bit more comprehensive
   // topNode should not have a sqlJoin entry...
   assert(topNode.sqlJoin == null)
 }
 
-export function base64(str) {
+export function objToCursor(obj) {
+  const str = JSON.stringify(obj)
   return new Buffer(str).toString('base64')
 }
 
-export function parseCursor(opaque) {
-  const clear = new Buffer(opaque, 'base64').toString()
-  const [ _, value ] = clear.split(':')
-  if (!_ || !value) {
-    throw new Error('invalid cursor format')
-  }
-  return value
+export function cursorToObj(cursor) {
+  const str = new Buffer(cursor, 'base64').toString()
+  return JSON.parse(str)
 }
+