+++ test/pg/keyset-paging.js
@@ -1,8 +1,9 @@
 import test from 'ava'
 import { graphql } from 'graphql'
-import schemaRelay from '../../example/schema-relay-paginate-1/index'
+import schemaRelay from '../../example/schema-relay-paginate-2/index'
 import { partial } from 'lodash'
-import { offsetToCursor, toGlobalId } from 'graphql-relay'
+import { toGlobalId } from 'graphql-relay'
+import { objToCursor } from '../../src/util'
 
 // monkey-patch the array prototype because these are tests and IDGAF
 Object.defineProperty(Array.prototype, 'last', {
@@ -25,7 +26,7 @@ function stringifyArgs(args) {
   return `(${argArr.join(', ')})`
 }
 
-const pageInfo = 'pageInfo { hasNextPage, startCursor, endCursor }'
+const pageInfo = 'pageInfo { hasNextPage, hasPreviousPage, startCursor, endCursor }'
 
 function makeUsersQuery(args) {
   let argString = stringifyArgs(args)
@@ -46,17 +47,13 @@ test('should handle pagination at the root', async t => {
   t.is(errors, undefined)
   t.deepEqual(data.users.pageInfo, {
     hasNextPage: false,
-    startCursor: offsetToCursor(0),
-    endCursor: offsetToCursor(4)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ id: 1 }),
+    endCursor: objToCursor({ id: 5 })
   })
-  t.deepEqual(data.users.edges[0], {
-    cursor: offsetToCursor(0),
-    node: {
-      id: toGlobalId('User', 1),
-      fullName: 'Alivia Waelchi',
-      email: 'Mohammed.Hayes@hotmail.com'
-    }
-  })
+  // generate globalIds for users 1 thru 5
+  const ids = Array.apply(null, Array(5)).map((_, i) => toGlobalId('User', i + 1))
+  t.deepEqual(data.users.edges.map(edge => edge.node.id), ids)
   t.is(data.users.edges.last().cursor, data.users.pageInfo.endCursor)
 })
 
@@ -66,11 +63,12 @@ test('should handle root pagination with "first" arg', async t => {
   t.is(errors, undefined)
   t.deepEqual(data.users.pageInfo, {
     hasNextPage: true,
-    startCursor: offsetToCursor(0),
-    endCursor: offsetToCursor(1)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ id: 1 }),
+    endCursor: objToCursor({ id: 2 })
   }, 'page info is accurate')
   t.deepEqual(data.users.edges[0], {
-    cursor: offsetToCursor(0),
+    cursor: objToCursor({ id: 1 }),
     node: {
       id: toGlobalId('User', 1),
       fullName: 'Alivia Waelchi',
@@ -81,16 +79,17 @@ test('should handle root pagination with "first" arg', async t => {
 })
 
 test('should handle root pagination with "first" and "after" args', async t => {
-  const query = makeUsersQuery({ first: 2, after: offsetToCursor(1) })
+  const query = makeUsersQuery({ first: 2, after: objToCursor({ id: 2 }) })
   const { data, errors } = await run(query)
   t.is(errors, undefined)
   t.deepEqual(data.users.pageInfo, {
     hasNextPage: true,
-    startCursor: offsetToCursor(2),
-    endCursor: offsetToCursor(3)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ id: 3 }),
+    endCursor: objToCursor({ id: 4 })
   }, 'page info is accurate')
   t.deepEqual(data.users.edges[0], {
-    cursor: offsetToCursor(2),
+    cursor: objToCursor({ id: 3 }),
     node: {
       id: toGlobalId('User', 3),
       fullName: 'Coleman Abernathy',
@@ -101,17 +100,18 @@ test('should handle root pagination with "first" and "after" args', async t => {
 })
 
 test('should handle the last page of root pagination', async t => {
-  const query = makeUsersQuery({ first: 2, after: offsetToCursor(3) })
+  const query = makeUsersQuery({ first: 2, after: objToCursor({ id: 4 }) })
   const { data, errors } = await run(query)
   t.is(errors, undefined)
   t.deepEqual(data.users.pageInfo, {
     hasNextPage: false,
-    startCursor: offsetToCursor(4),
-    endCursor: offsetToCursor(4)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ id: 5 }),
+    endCursor: objToCursor({ id: 5 })
   }, 'page info is accurate')
   t.is(data.users.edges.length, 1)
   t.deepEqual(data.users.edges[0], {
-    cursor: offsetToCursor(4),
+    cursor: objToCursor({ id: 5 }),
     node: {
       id: toGlobalId('User', 5),
       fullName: 'Ocie Ruecker',
@@ -122,12 +122,13 @@ test('should handle the last page of root pagination', async t => {
 })
 
 test('should return nothing after the end of root pagination', async t => {
-  const query = makeUsersQuery({ first: 3, after: offsetToCursor(4) })
+  const query = makeUsersQuery({ first: 3, after: objToCursor({ id: 5 }) })
   const { data, errors } = await run(query)
   t.is(errors, undefined)
   t.deepEqual(data.users, {
     pageInfo: {
       hasNextPage: false,
+      hasPreviousPage: false,
       startCursor: null,
       endCursor: null
     },
@@ -135,6 +136,32 @@ test('should return nothing after the end of root pagination', async t => {
   })
 })
 
+test('should handle backward pagination at root with "last" arg', async t => {
+  const query = makeUsersQuery({ last: 2 })
+  const { data, errors } = await run(query)
+  t.is(errors, undefined)
+  t.deepEqual(data.users.pageInfo, {
+    hasNextPage: false,
+    hasPreviousPage: true,
+    startCursor: objToCursor({ id: 4 }),
+    endCursor: objToCursor({ id: 5 })
+  })
+  t.is(data.users.edges[0].node.id, toGlobalId('User', 4))
+  t.is(data.users.edges[1].node.id, toGlobalId('User', 5))
+})
+
+test('should handle backward pagination at root with "last" and "before" args', async t => {
+  const query = makeUsersQuery({ last: 1, before: objToCursor({ id: 2 }) })
+  const { data, errors } = await run(query)
+  t.is(errors, undefined)
+  t.deepEqual(data.users.pageInfo, {
+    hasNextPage: false,
+    hasPreviousPage: false,
+    startCursor: objToCursor({ id: 1 }),
+    endCursor: objToCursor({ id: 1 })
+  })
+  t.is(data.users.edges.length, 1)
+})
 
 function makePostsQuery(args) {
   let argString = stringifyArgs(args)
@@ -158,12 +185,13 @@ test('should handle pagination in a nested field', async t => {
   const posts = data.user.posts
   t.deepEqual(posts.pageInfo, {
     hasNextPage: false,
-    startCursor: offsetToCursor(0),
-    endCursor: offsetToCursor(7)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ created_at: '2016-04-17T18:49:15.942Z', id: 2 }),
+    endCursor: objToCursor({ created_at: '2015-11-15T08:26:11.331Z', id: 30 })
   })
   t.is(posts.edges.length, 8)
   t.deepEqual(posts.edges[0], {
-    cursor: offsetToCursor(0),
+    cursor: objToCursor({ created_at: '2016-04-17T18:49:15.942Z', id: 2 }),
     node: {
       id: toGlobalId('Post', 2),
       body: 'Adipisci voluptate laborum minima sunt facilis sint quibusdam ut. Deserunt nemo pariatur sed facere accusantium quis. Nobis aut voluptate inventore quidem explicabo.'
@@ -179,26 +207,29 @@ test('nested paging should handle "first" arg', async t => {
   const posts = data.user.posts
   t.deepEqual(posts.pageInfo, {
     hasNextPage: true,
-    startCursor: offsetToCursor(0),
-    endCursor: offsetToCursor(2)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ created_at: '2016-04-17T18:49:15.942Z', id: 2 }),
+    endCursor: objToCursor({ created_at: '2016-04-13T15:07:15.119Z', id: 33 })
   })
   t.is(posts.edges.length, 3)
   t.is(posts.edges.last().cursor, posts.pageInfo.endCursor)
 })
 
-test('nested paging should handle "first" and "after" args that reaches the last page', async t => {
-  const query = makePostsQuery({ first: 5, after: offsetToCursor(3) })
+test('nested paging should handle "last" and "before" args', async t => {
+  const query = makePostsQuery({ last: 2, before: objToCursor({ created_at: '2016-04-13T15:07:15.119Z', id: 33 }) })
   const { data, errors } = await run(query)
   t.is(errors, undefined)
-  const posts = data.user.posts
-  t.deepEqual(posts.pageInfo, {
+  const expect = {
     hasNextPage: false,
-    startCursor: offsetToCursor(4),
-    endCursor: offsetToCursor(7)
-  })
-  t.is(posts.edges.length, 4)
-  t.is(posts.edges.last().cursor, posts.pageInfo.endCursor)
+    hasPreviousPage: false,
+    startCursor: objToCursor({ created_at: '2016-04-17T18:49:15.942Z', id: 2 }),
+    endCursor: objToCursor({ created_at: '2016-04-15T03:29:31.212Z', id: 28 })
+  }
+  t.deepEqual(data.user.posts.pageInfo, expect)
+  t.is(data.user.posts.edges[0].node.id, toGlobalId('Post', 2))
+  t.is(data.user.posts.edges[1].node.id, toGlobalId('Post', 28))
 })
+
 
 test('can handle nested pagination', async t => {
   const query = `{
@@ -221,31 +252,6 @@ test('can handle nested pagination', async t => {
   t.is(data.users.edges[0].node.fullName, 'Alivia Waelchi')
   t.is(data.users.edges[0].node.posts.edges.length, 2)
   t.is(data.users.edges[0].node.posts.edges[0].node.body, 'Adipisci voluptate laborum minima sunt facilis sint quibusdam ut. Deserunt nemo pariatur sed facere accusantium quis. Nobis aut voluptate inventore quidem explicabo.')
-})
-
-test('can go to each second page in a nested connection', async t => {
-  const query = `{
-    users(first: 2) {
-      edges {
-        node {
-          id
-          fullName
-          posts(first: 2, after: "${offsetToCursor(1)}") {
-            edges {
-              cursor
-              node { id, body }
-            }
-          }
-        }
-      }
-    }
-  }`
-  const { data, errors } = await run(query)
-  t.is(errors, undefined)
-  t.is(data.users.edges[0].node.id, toGlobalId('User', 1))
-  t.deepEqual(data.users.edges[0].node.posts.edges.map(edge => edge.node.id), [ toGlobalId('Post', 33), toGlobalId('Post', 38)])
-  t.is(data.users.edges[1].node.id, toGlobalId('User', 2))
-  t.deepEqual(data.users.edges[1].node.posts.edges.map(edge => edge.node.id), [ toGlobalId('Post', 1), toGlobalId('Post', 50)])
 })
 
 test('can handle deeply nested pagination', async t => {
@@ -281,17 +287,18 @@ test('can handle deeply nested pagination', async t => {
   const { data, errors } = await run (query)
   t.is(errors, undefined)
   const comments = data.users.edges[0].node.posts.edges[0].node.comments
-  t.deepEqual(comments.pageInfo, {
+  const expect = {
     hasNextPage: true,
-    startCursor: offsetToCursor(0),
-    endCursor: offsetToCursor(2)
-  })
+    startCursor: objToCursor({ id: 233 }),
+    endCursor: objToCursor({ id: 116 })
+  }
+  t.deepEqual(comments.pageInfo, expect)
   t.is(comments.edges.length, 3)
   t.deepEqual(comments.edges[0], {
-    cursor: offsetToCursor(0),
+    cursor: objToCursor({ id: 233 }),
     node: {
-      id: toGlobalId('Comment', 18),
-      body: 'bypassing the hard drive won\'t do anything, we need to back up the primary EXE bandwidth!',
+      id: toGlobalId('Comment', 233),
+      body: 'I\'ll reboot the digital SCSI system, that should bus the USB protocol!',
       author: {
         fullName: 'Coleman Abernathy'
       }
@@ -303,7 +310,7 @@ test('can handle deeply nested pagination', async t => {
 test('handle a conection type with a many-to-many', async t => {
   const query = `{
     user(id: 2) {
-      following(first: 2, after:"${offsetToCursor(0)}") {
+      following(first: 2, after: "${objToCursor({ created_at: '2016-01-01T16:28:00.051Z', followee_id: 1 })}") {
         pageInfo {
           hasNextPage
           startCursor
@@ -322,12 +329,12 @@ test('handle a conection type with a many-to-many', async t => {
   t.is(errors, undefined)
   t.deepEqual(data.user.following.pageInfo, {
     hasNextPage: true,
-    startCursor: offsetToCursor(1),
-    endCursor: offsetToCursor(2)
+    startCursor: objToCursor({ created_at: '2016-05-18T21:35:54.601Z', followee_id: 3 }),
+    endCursor: objToCursor({ created_at: '2016-06-15T08:56:18.519Z', followee_id: 2 })
   })
   t.deepEqual(data.user.following.edges, [
-    { node: { id: toGlobalId('User', 2), fullName: 'Hudson Hyatt' } },
-    { node: { id: toGlobalId('User', 3), fullName: 'Coleman Abernathy' } }
+    { node: { id: toGlobalId('User', 3), fullName: 'Coleman Abernathy' } },
+    { node: { id: toGlobalId('User', 2), fullName: 'Hudson Hyatt' } }
   ])
 })